Rearrange screen composables
